diag_pf_info proposal
This file currently contains:

$form = new Form(false);

which creates a fatal error because Form() takes a Form_Button, not a
boolean. I believe it should be:

$form = new Form(new Form_Button(
	'',
	''
));

In my opinion, however, a whole panel for one checkbox is a waste of
valuable screen space, particularly since the original pfSense page did
not have that checkbox.

Please consider this version with a minimal checkbox.